Fixed narrowing errors from clang compilation.
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/trunk@651 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ src/course_render.cpp
@@ -154,10 +154,19 @@ void DrawTrees() {
 		};
 
 		const GLfloat vtx[] = {
-			-itemRadius*normal.z, 0.f,        itemRadius*normal.x,
-			itemRadius*normal.z,  0.f,        -itemRadius*normal.x,
-			itemRadius*normal.z,  itemHeight, -itemRadius*normal.x,
-			-itemRadius*normal.z, itemHeight, itemRadius*normal.x
+                  static_cast<GLfloat>(-itemRadius*normal.z),
+                  0.f,
+                  static_cast<GLfloat>(itemRadius*normal.x),
+
+		  static_cast<GLfloat>(itemRadius*normal.z),
+                  0.f,
+                  static_cast<GLfloat>(-itemRadius*normal.x),
+		  static_cast<GLfloat>(itemRadius*normal.z),
+                  static_cast<GLfloat>(itemHeight),
+                  static_cast<GLfloat>(-itemRadius*normal.x),
+		  static_cast<GLfloat>(-itemRadius*normal.z),
+                  static_cast<GLfloat>(itemHeight),
+                  static_cast<GLfloat>(itemRadius*normal.x)
 		};
 		glEnableClientState(GL_VERTEX_ARRAY);
 		glEnableClientState(GL_TEXTURE_COORD_ARRAY);
